refactor: split lib/chat-turn.ts into shell-tokenizer and content-persistence
